chore: restructure user, auth services #11
