--- src/dummy/Dummy.txt
@@ -1,40 +1,0 @@
-
-<!-- Normal Mode Specific Content -->
-      <!-- <div
-        v-if="selectedGameMode.id === 'normal'"
-        class="normal-mode-options"
-      >
-        <h3>난이도 선택</h3>
-        <div class="level-selector">
-          <div
-            v-for="level in levels"
-            :key="level.value"
-            :class="[
-              'level-card',
-              { selected: selectedLevel === level.value },
-            ]"
-            @click="selectLevel(level.value)"
-          >
-            <h4>{{ level.title }}</h4>
-            <p>{{ level.description }}</p>
-          </div>
-        </div>
-      
-      </div> -->
-
-
-
-      {
-    id: "normal",
-    title: "일반 게임",
-    icon: "fas fa-play-circle",
-    shortDescription: "다양한 난이도로 실력을 키워보세요",
-    fullDescription:
-      "점진적으로 난이도를 높여가며 한국 전역의 로드뷰 실력을 테스트할 수 있는 모드입니다.",
-  },
-
-  const levels = [
-  { value: 1, title: "Level 1", description: "쉬운 난이도, 기본 로드뷰" },
-  { value: 2, title: "Level 2", description: "중간 난이도, 도전적인 로드뷰" },
-  { value: 3, title: "Level 3", description: "어려운 난이도, 전문가 수준" },
-];
