+++ src/features/auth/services/auth.service.js
@@ -1,4 +1,4 @@
-import apiClient from '../apiClient';
+import apiClient from 'src/core/api/client.js'
 
 /**
  * 인증 관련 API 서비스
